[turbopack] Move WellKnown*Kind enums into well_known/kinds.rs
